Refactor: Side User Profile Card-StickyLayout과 결합
- 불필요한 혹은 중복된 스타일 class 제거
- useDeviceType Hooks 추가
- TABLET_BREAKPOINT 상수 추가
