+++ app/_components/side-sticky-layout/index.tsx
@@ -1,24 +1,16 @@
 'use client';
 
 import type { PropsWithChildren } from 'react';
-import { useEffect, useState } from 'react';
 
-import useWindowSize from '@/_hooks/useWindowSize';
+import useDeviceType from '@/_hooks/useDeviceType';
+
+import SideUserProfileCard from '../side-user-profile-card';
 
 /**
  * 왼쪽 내정보 사이드 메뉴 고정된 레이아웃
  */
 function StickyLayout({ children }: PropsWithChildren) {
-  const windowSize = useWindowSize();
-  const [isLoading, setIsLoading] = useState(true);
-
-  useEffect(() => {
-    if (windowSize !== null) {
-      setIsLoading(false);
-    }
-  }, [windowSize]);
-
-  const isPC = windowSize > 768;
+  const { isDevice: isPC, isLoading } = useDeviceType();
 
   return (
     <div className="flex gap-6">
@@ -30,7 +22,11 @@ function StickyLayout({ children }: PropsWithChildren) {
           </div>
         </section>
       ) : (
-        isPC && <section className="sticky top-20 h-fit max-w-[384px] flex-1 rounded-xl bg-slate-300 p-6">사이드 스티키 메뉴</section>
+        isPC && (
+          <section className="sticky top-20 h-fit max-w-[384px] flex-1">
+            <SideUserProfileCard />
+          </section>
+        )
       )}
       <section className="min-w-[430px] flex-1">{children}</section>
     </div>
