+++ app/_constants/deviceBreakpoint.ts
@@ -1,0 +1,1 @@
+export const TABLET_BREAKPOINT = 768;
