+++ app/_hooks/useWindowSize.ts
@@ -1,6 +1,9 @@
 import { useEffect, useState } from 'react';
 import { debounce } from 'lodash';
 
+/**
+ * 브라우저 창의 너비(width)를 추적하고 반환합니다.
+ */
 const useWindowSize = () => {
   const [windowSize, setWindowSize] = useState(0);
