+++ app/page.tsx
@@ -1,10 +1,11 @@
 import NoticeCalender from './_components/calendar-notice';
+import StickyLayout from './_components/side-sticky-layout';
 import SideUserProfileCard from './_components/side-user-profile-card';
 
 export default function Home() {
   return (
-    <main className="flex min-h-screen flex-col items-center justify-between p-24">
-      <SideUserProfileCard />
+    <main>
+      <StickyLayout />
       <a href="/my">마이페이지 이동</a>
     </main>
   );
